fix subdivide mesh by facecolor
